# Improve `Value.Emit()` with strconv.