--- a/api/core/key_test.go
+++ b/api/core/key_test.go
@@ -203,3 +203,43 @@ func TestEmit(t *testing.T) {
 		})
 	}
 }
+
+func BenchmarkEmitBool(b *testing.B) {
+	b.ReportAllocs()
+	for i := 0; i < b.N; i++ {
+		n := core.Bool(i%2 == 0)
+		_ = n.Emit()
+	}
+}
+
+func BenchmarkEmitInt64(b *testing.B) {
+	b.ReportAllocs()
+	for i := 0; i < b.N; i++ {
+		n := core.Int64(int64(i))
+		_ = n.Emit()
+	}
+}
+
+func BenchmarkEmitUInt64(b *testing.B) {
+	b.ReportAllocs()
+	for i := 0; i < b.N; i++ {
+		n := core.Uint64(uint64(i))
+		_ = n.Emit()
+	}
+}
+
+func BenchmarkEmitFloat64(b *testing.B) {
+	b.ReportAllocs()
+	for i := 0; i < b.N; i++ {
+		n := core.Float64(float64(i))
+		_ = n.Emit()
+	}
+}
+
+func BenchmarkEmitFloat32(b *testing.B) {
+	b.ReportAllocs()
+	for i := 0; i < b.N; i++ {
+		n := core.Float32(float32(i))
+		_ = n.Emit()
+	}
+}