`strconv` methods has a better performance than using `fmt.Sprint()`. 

*Floats didn't show significant improvement and it actually allocates more bytes. It still uses `fmt.Sprint()`.*

benchstat output:
```
name           old time/op    new time/op    delta
EmitBool-6        115ns ±12%       3ns ± 7%   -97.02%  (p=0.008 n=5+5)
EmitInt64-6       156ns ± 4%      74ns ± 4%   -52.90%  (p=0.008 n=5+5)
EmitUInt64-6      164ns ±10%      76ns ±15%   -53.70%  (p=0.008 n=5+5)
EmitFloat64-6     221ns ± 7%     200ns ±22%      ~     (p=0.421 n=5+5)
EmitFloat32-6     213ns ± 4%     223ns ±12%      ~     (p=0.452 n=5+5)

name           old alloc/op   new alloc/op   delta
EmitBool-6        8.00B ± 0%     0.00B       -100.00%  (p=0.008 n=5+5)
EmitInt64-6       16.0B ± 0%      7.0B ± 0%   -56.25%  (p=0.008 n=5+5)
EmitUInt64-6      16.0B ± 0%      7.0B ± 0%   -56.25%  (p=0.008 n=5+5)
EmitFloat64-6     22.0B ± 0%     39.0B ± 0%   +77.27%  (p=0.008 n=5+5)
EmitFloat32-6     15.0B ± 0%     39.0B ± 0%  +160.00%  (p=0.008 n=5+5)

name           old allocs/op  new allocs/op  delta
EmitBool-6         1.00 ± 0%      0.00       -100.00%  (p=0.008 n=5+5)
EmitInt64-6        2.00 ± 0%      0.00       -100.00%  (p=0.008 n=5+5)
EmitUInt64-6       2.00 ± 0%      0.00       -100.00%  (p=0.008 n=5+5)
EmitFloat64-6      2.00 ± 0%      2.00 ± 0%      ~     (all equal)
EmitFloat32-6      2.00 ± 0%      2.00 ± 0%      ~     (all equal)

```